requirements.txt
--index-url https://pypi.python.org/simple/
-e .

PyYAML==3.11
